projeler/haber-bot/requirements.txt
feedparser==6.0.11
deep-translator==1.11.4
python-dotenv==1.0.1
schedule==1.2.2
requests==2.32.3
